Improve fighting.

+++ characters/actions/efightaction.cpp
@@ -17,7 +17,7 @@ eFightAction::eFightAction(eCharacter* const c,
 
 void eFightAction::increment(const int by) {
     const auto c = character();
-    const int a = c->attack();
+    const double a = c->attack();
     const bool dead = mOpponent ? mOpponent->defend(by*a) : true;
     if(dead || c->dead()) {
         setState(eCharacterActionState::finished);
